Move the scalars and tests out of mutable/immutable directories
