+++ src/Scalars/ArrayOfBoolean.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace ArrayOf\Scalars\Mutable;
+namespace ArrayOf\Scalars;
 
 use ArrayOf\AbstractArrayOf;
 
-final class ArrayOfString extends AbstractArrayOf
+final class ArrayOfBoolean extends AbstractArrayOf
 {
     protected function typeToEnforce(): string
     {
-        return self::SCALAR_STRING;
+        return self::SCALAR_BOOLEAN;
     }
 }
