+++ src/Scalars/ArrayOfString.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace ArrayOf\Scalars\Mutable;
+namespace ArrayOf\Scalars;
 
 use ArrayOf\AbstractArrayOf;
 
-final class ArrayOfFloat extends AbstractArrayOf
+final class ArrayOfString extends AbstractArrayOf
 {
     protected function typeToEnforce(): string
     {
-        return self::SCALAR_DOUBLE;
+        return self::SCALAR_STRING;
     }
 }
