+++ tests/Unit/Scalars/ArrayOfBooleanTest.php
@@ -2,21 +2,21 @@
 
 declare(strict_types=1);
 
-namespace ArrayOfTest\Unit\Scalars\Mutable;
+namespace ArrayOfTest\Unit\Scalars;
 
 use ArrayOf\AbstractArrayOf;
 use ArrayOf\Exceptions\InvalidInstantiationType;
-use ArrayOf\Scalars\Mutable\ArrayOfString;
+use ArrayOf\Scalars\ArrayOfBoolean;
 use Generator;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 
-final class ArrayOfStringTest extends TestCase
+final class ArrayOfBooleanTest extends TestCase
 {
     public function testConstruct(): void
     {
-        $test = new ArrayOfString(['test']);
-        self::assertInstanceOf(ArrayOfString::class, $test);
+        $test = new ArrayOfBoolean([true]);
+        self::assertInstanceOf(ArrayOfBoolean::class, $test);
         self::assertInstanceOf(AbstractArrayOf::class, $test);
     }
 
@@ -26,7 +26,7 @@ final class ArrayOfStringTest extends TestCase
     public function testInvalidScalarInputType(array $arguments): void
     {
         $this->expectException(InvalidInstantiationType::class);
-        new ArrayOfString($arguments);
+        new ArrayOfBoolean($arguments);
     }
 
     public function providerInvalidScalarInputType(): Generator
@@ -39,12 +39,12 @@ final class ArrayOfStringTest extends TestCase
             'arguments' => [1.23, 4.56],
         ];
 
-        yield 'Receiving booleans' => [
-            'arguments' => [true, false],
-        ];
-
         yield 'Receiving stdClasses' => [
             'arguments' => [new stdClass(), new stdClass()],
+        ];
+
+        yield 'Receiving strings' => [
+            'arguments' => ['str1', 'str2'],
         ];
 
         yield 'Receiving a mix of all scalars' => [
