+++ tests/Unit/Scalars/ArrayOfIntegerTest.php
@@ -2,21 +2,21 @@
 
 declare(strict_types=1);
 
-namespace ArrayOfTest\Unit\Scalars\Mutable;
+namespace ArrayOfTest\Unit\Scalars;
 
 use ArrayOf\AbstractArrayOf;
 use ArrayOf\Exceptions\InvalidInstantiationType;
-use ArrayOf\Scalars\Mutable\ArrayOfFloat;
+use ArrayOf\Scalars\ArrayOfInteger;
 use Generator;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 
-final class ArrayOfFloatTest extends TestCase
+final class ArrayOfIntegerTest extends TestCase
 {
     public function testConstruct(): void
     {
-        $test = new ArrayOfFloat([1.5]);
-        self::assertInstanceOf(ArrayOfFloat::class, $test);
+        $test = new ArrayOfInteger([1]);
+        self::assertInstanceOf(ArrayOfInteger::class, $test);
         self::assertInstanceOf(AbstractArrayOf::class, $test);
     }
 
@@ -26,13 +26,13 @@ final class ArrayOfFloatTest extends TestCase
     public function testInvalidScalarInputType(array $arguments): void
     {
         $this->expectException(InvalidInstantiationType::class);
-        new ArrayOfFloat($arguments);
+        new ArrayOfInteger($arguments);
     }
 
     public function providerInvalidScalarInputType(): Generator
     {
-        yield 'Receiving integers' => [
-            'arguments' => [1, 2],
+        yield 'Receiving floats' => [
+            'arguments' => [1.23, 4.56],
         ];
 
         yield 'Receiving booleans' => [
